requirements.txt
openai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
gspread
streamlit

accelerate >= 0.12.0
seqeval
datasets >= 1.8.0
torch >= 1.3
evaluate
transformers